About button update
Took off the background from the icon and changed the icon to green for light mode and yellow for dark mode. Attempted to make the icon smaller on mobile but had trouble viewing the mobile version of the app so I am not sure how the change I made currently looks on mobile.