Some samples for custom scrollbars to match theme
I could not figure out and get it to work using Tailwind CSS, so I created a separate "styles.css file" in the public folder and used the "index.html" file in public folder as a template for the designs. You can see it in the browser with using the VS Code extension live server.